test: RecipeRepositoryImpl 테스트 수정
